Run pygmt plotting in seperate subprocess to better handle sessions
